// src/app/main/apps/procesos/contratacion/models/contratacion.model.ts
export class EtapaResponseModel {
    idEtapa: number;
    idProceso: number;
    anio: number;
    idEtapaFase?: number;
    codigoProceso: string;
    descripcionEtapa: string;
    descripcionRegimenLaboral: string;
    descripcionEtapaFase: string;
    descripcionNumeroConvocatoria: string;
    descripcionTipoProceso: string;
    fechaCreacion: string;
    descripcionNivelEducativo?: string;
    descripcionCargosConvocados?: string;
    idRegimenLaboral?: number;
    idNivelEducativo?: number;
    codigoRegimenLaboral?: number;
    codigoEtapaFase: number;
    codigoEtapa: string;
    descripcioncargosconvocados?: string;
    descripcionEstadoEtapaProceso: string;
    idEstadoDesarrollo?: number;
    // adicionales
    accesoPublicacionPlaza?: boolean;
    accesoValidarPlazas?: boolean;
    accesoAprobarPlazas?: boolean;
    accesoRechazarPlazas?: boolean;
    accesoPublicarPlazas?: boolean;
    accesoPostulantes?: boolean;
    accesoCargaMasivaPostulantes?: boolean;
    accesoEliminarCargaPostulantes?: boolean;
    accesoCalificacion?: boolean;
    accesoPublicarCalificacion?: boolean;
    accesoCargaMasivaCalificacion?: boolean;
    accesoEliminarCargaCalificacion?: boolean;
    accesoCalificacionAutomatica?: boolean;
    accesoAdjudicacion?: boolean;
    accesoPublicarAdjudicados?: boolean;
    accesoCargaMasivaAdjudicacion?: boolean;
    accesoEliminarCargaAdjudicacion?: boolean;
    accesoConsolidado?: boolean;
    accesoAprobarPlazaConsolidado?: boolean;
    accesoPlazaPrepublicada?: boolean;
    accesoPlazaResultadoFinal?: boolean;
    accesoRegistrarCalificacion?: boolean;
    accesoRegistrarReclamoCalificacion?: boolean;
    accesoFinalizarEtapa?: boolean;
    accesoAdjudicarPlaza?: boolean;
    accesoNoAdjudicarPlaza?: boolean;
}
// ...
export class CalificacionModel {
    idCalificacion: number;
    idProceso: number;
    idEtapa: number;
    idPostulacion: number;
    idEstadoCalificacion: number;
    publicadoPreliminar: boolean;
    codigoCargaMasiva: string;
    detalleReclamo: string;
    noSePresento: boolean;
    ordenMerito: number;
    puntajeTotal: number;
    usuarioRegistro: string;
    requisitosMinimos: DetalleCalificacionModel[] = [];
    formacionesAcademicas: DetalleCalificacionModel[] = [];
    formacionesProfesionales: DetalleCalificacionModel[] = [];
    experienciasLaborales: DetalleCalificacionModel[] = [];
}

export class DetalleCalificacionModel {
    idDetalleCalificacion: number;
    idTipoCalificacion: number;
    idCriterioCalificacion: number;
    idTipoPuntaje: number;
    idDetalleCalificacionPadre: number;
    puntajeUnitario: number;
    cantidad: number;
    puntaje: number;
    acreditaDocumento: number;
}
// ...
export class PlazaModel {
    registro: number;
    totalRegistro: number;
    idPlaza: number;
    codigoModular: string;
    descripcionInstitucionEducativa: string;
    abreviaturaModalidadEducativa: string;
    descripcionNivelEducativo: string;
    descripcionTipoGestionEducativa: string;
    nombreZona: string;
    eib: string;
    codigoPlaza: string;
    descripcionCargo: string;
    descripcionAreaCurricular: string;
    jornadaLaboral: number;
    descripcionTipoPlaza: string;
    descripcionMotivoVacancia: string;
    vigenciaInicio: string;
    vigenciaFin: string;
}
// ...
export class ServidorPublicoModel {
    registro: number;
    totalRegistro: number;
    idServidorPublico: number;
    idPersona: number;
    idTipoDocumentoIdentidad: number;
    numeroDocumentoIdentidad: string;
    numeroDocumentoIdentidadCompleto: string;
    nombreCompleto: string;
    fechaNacimiento: string;
    descripcionUgel: string;
    centroTrabajo: string;
    abreviaturaRegimenLaboral: string;
    situacionLaboral: string;
    idSituacionLaboral: number;
    descripcionGenero: string;
    descripcionNacionalidad: string;
    primerApellido: string;
    segundoApellido: string;
    nombres: string;
}
// ...
export class CalificacionResponse {
    idPostulacion: number;
    idProceso: number;
    idEtapa: number;
    idPersona: number;
    idTipoDocumentoIdentidad: number;
    numeroDocumentoIdentidad: string;
    numeroDocumentoIdentidadCompleto: string;
    fechaNacimiento: string;
    primerApellido: string;
    segundoApellido: string;
    nombres: string;
    descripcionGenero: string;
    descripcionModalidadEducativa: string;
    descripcionNivelEducativo: string;
    descripcionAreaCurricular: string;
    descripcionCargo: string;
    jornadaLaboral: number;
    descripcionRegimenLaboral: string;
    idCalificacion: number;
    descripcionGrupoInscripcion: string;
    puntajeTotal: number;
    puntajeFormacionAcademica: number;
    puntajeFormacionProfesional: number;
    puntajeExperienciaLaboral: number;
    puntajeEvaluacionExpediente: number;
    puntajeDesempate: number;
    publicadoPreliminar: boolean;
    noSePresento: boolean;
    empate: boolean;
}